+++ source/utils/errors-ids.js
@@ -1,0 +1,10 @@
+/**
+ * @license
+ * Copyright (C) 2020 Pryv S.A. https://pryv.com - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential
+ */
+module.exports = {
+  ItemAlreadyExists: 'item-already-exists',
+  InvalidInvitationToken: 'invitationToken-invalid'
+}