+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGES in 1.X
 
+### 1.6.0
+
+ - POST /users system endpoint to register new user only in service-register
+ - PUT /users system endpoint to update a user only in service-register
+ - POST /users/validate system endpoint to validate user data only in service-register
+
 ## 1.3.X
 
 ### 1.3.65
@@ -8,7 +14,7 @@
 
 ### 1.3.64
 
- - GET /admin/users/:username enpoint
+ - GET /admin/users/:username endpoint
  - /admin/users/invitations renamed to /admin/invitations
 
 ### 1.3.62